+++ conanfile.py
@@ -9,7 +9,7 @@ required_conan_version = ">=1.60.0"
 
 class HomestoreConan(ConanFile):
     name = "homestore"
-    version = "6.4.61"
+    version = "6.4.62"
 
     homepage = "https://github.com/eBay/Homestore"
     description = "HomeStore Storage Engine"
