+++ src/include/homestore/replication/repl_decls.h
@@ -19,17 +19,18 @@ namespace homestore {
 VENUM(ReplServiceError, int32_t,
       OK = 0,         // Everything OK
       CANCELLED = -1, // Request was cancelled
-      TIMEOUT = -2, 
-      NOT_LEADER = -3, 
-      BAD_REQUEST = -4, 
-      SERVER_ALREADY_EXISTS = -5, 
+      TIMEOUT = -2,
+      NOT_LEADER = -3,
+      BAD_REQUEST = -4,
+      SERVER_ALREADY_EXISTS = -5,
       CONFIG_CHANGING = -6,
-      SERVER_IS_JOINING = -7, 
-      SERVER_NOT_FOUND = -8, 
-      CANNOT_REMOVE_LEADER = -9, 
+      SERVER_IS_JOINING = -7,
+      SERVER_NOT_FOUND = -8,
+      CANNOT_REMOVE_LEADER = -9,
       SERVER_IS_LEAVING = -10,
-      TERM_MISMATCH = -11, 
-      RESULT_NOT_EXIST_YET = -10000, 
+      TERM_MISMATCH = -11,
+      RETRY_REQUEST = -12,
+      RESULT_NOT_EXIST_YET = -10000,
       NOT_IMPLEMENTED = -10001,
       NO_SPACE_LEFT = -20000,
       DRIVE_WRITE_ERROR = -20001,
