+++ src/include/homestore/replication/repl_dev.h
@@ -42,7 +42,8 @@ VENUM(repl_req_state_t, uint32_t,
 VENUM(journal_type_t, uint16_t,
       HS_DATA_LINKED = 0,  // Linked data where each entry will store physical blkid where data reside
       HS_DATA_INLINED = 1, // Data is inlined in the header of journal entry
-      HS_CTRL_DESTROY = 2  // Control message to destroy the repl_dev
+      HS_CTRL_DESTROY = 2, // Control message to destroy the repl_dev
+      HS_CTRL_REPLACE = 3, // Control message to replace a member
 )
 
 struct repl_key {
@@ -345,6 +346,9 @@ public:
     /// However, it is expected that this call be idempotent. It is possible in rare scenarios that this can be called
     /// after restart in case crash happened during the destroy.
     virtual void on_destroy() = 0;
+
+    /// @brief Called when replace member is performed.
+    virtual void replace_member(replica_id_t member_out, replica_id_t member_in) = 0;
 
     /// @brief Called when the snapshot is being created by nuraft
     virtual AsyncReplResult<> create_snapshot(shared< snapshot_context > context) = 0;
