+++ src/tests/CMakeLists.txt
@@ -118,6 +118,10 @@ if (${io_tests})
     target_sources(test_raft_repl_dev PRIVATE test_raft_repl_dev.cpp)
     target_link_libraries(test_raft_repl_dev homestore ${COMMON_TEST_DEPS} GTest::gmock)
 
+    add_executable(test_raft_repl_dev_dynamic)
+    target_sources(test_raft_repl_dev_dynamic PRIVATE test_raft_repl_dev_dynamic.cpp)
+    target_link_libraries(test_raft_repl_dev_dynamic homestore ${COMMON_TEST_DEPS} GTest::gmock)
+
     can_build_epoll_io_tests(epoll_tests)
     if(${epoll_tests})
         add_test(NAME LogDev-Epoll COMMAND test_log_dev)
@@ -126,6 +130,7 @@ if (${io_tests})
         add_test(NAME MetaBlkMgr-Epoll COMMAND test_meta_blk_mgr)
         add_test(NAME DataService-Epoll COMMAND test_data_service)
         add_test(NAME RaftReplDev-Epoll COMMAND test_raft_repl_dev)
+        add_test(NAME RaftReplDevDynamic-Epoll COMMAND test_raft_repl_dev_dynamic)
         # add_test(NAME SoloReplDev-Epoll COMMAND test_solo_repl_dev)
     endif()
 
@@ -138,6 +143,7 @@ if (${io_tests})
         add_test(NAME SoloReplDev-Spdk COMMAND test_solo_repl_dev -- --spdk "true")
         add_test(NAME HomeRaftLogStore-Spdk COMMAND test_home_raft_logstore -- --spdk "true")
         add_test(NAME RaftReplDev-Spdk COMMAND test_raft_repl_dev -- --spdk "true")
+        add_test(NAME RaftReplDevDynamic-Spdk COMMAND test_raft_repl_dev_dynamic -- --spdk "true")
         if(${epoll_tests})
         SET_TESTS_PROPERTIES(MetaBlkMgr-Spdk PROPERTIES DEPENDS LogStore-Spdk)
         SET_TESTS_PROPERTIES(DataService-Spdk PROPERTIES DEPENDS MetaBlkMgr-Spdk)
