+++ src/tests/test_solo_repl_dev.cpp
@@ -136,6 +136,7 @@ public:
                       cintrusive< repl_req_ctx >& ctx) override {
             LOGINFO("Received error={} on repl_dev", enum_name(error));
         }
+        void replace_member(replica_id_t member_out, replica_id_t member_in) override {}
         void on_destroy() override {}
     };
